Wrap package in a project directory
This patch wraps the package generated by the `build_yaml` or
`build_json` functions in a project directory instead of writing the
files directly in the destination directory.

The project directory takes the same name as the package.

The `dump()` function was refactored to leverage `pathlib.Path`
features.

Errors reported by `bandit` were skipped as they were out of the scope
of this PR and should be fixed in a subsequent PR.

Drive-by:
* Updates `.gitignore`.
* Updates Azure pipelines to use `bandit` directly instead of the pytest
  plugin.