Automate GitHub releases
Automates the GitHub releases by leveraging the Keep a changelog format.
The source and wheel packages are also attached to the release and
uploaded to the PyPI.

This also builds the documentation strictly by treating warning as
errors, and run the examples to ensure none of them gets broken.

Fixes jdkandersson/OpenAlchemy#194