Refresh documentation
This is just an idea, so I did not push the research too far, but the
documentation starts to have **A LOT** of information, and the users
have to scroll a lot up and down, so I thought we could think of
refreshing the documentation a little.

As I started to research a new theme for my own projects, I found
[Furo](https://pradyunsg.me/furo/quickstart/) which really looks
awesome!

I think it brings the following benefits:

* It REALLY looks good!
* The side bars (left and right), make the navigation through the
  different sections clearer and quicker.
* We could easily add a "development" section in the left side bar to
  emphasize the contribultion guidelines, changelog, etc.

If we decide to move forward with this idea, there would be some
rewrite/reorganiztion to do, but nothing too crazy.

Drive-bys:
* Fixes ToC errors.
* Uses sphinx-autobuild.
* Reorganizes content.